Added the option that simulate a callback from Camera Intent for the UI tests and update the BikeTest
