+++ app/src/androidTest/java/app/igormatos/botaprarodar/bicycle/BicycleRobot.kt
@@ -5,15 +5,20 @@ import android.app.Instrumentation
 import android.content.Intent
 import android.graphics.BitmapFactory
 import android.os.Bundle
+import android.provider.MediaStore
+import androidx.test.core.app.ActivityScenario
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
+import androidx.test.espresso.intent.Intents.intending
+import androidx.test.espresso.intent.matcher.IntentMatchers.hasAction
 import androidx.test.espresso.intent.rule.IntentsTestRule
-import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
-import androidx.test.espresso.matcher.ViewMatchers.withId
+import androidx.test.espresso.matcher.ViewMatchers.*
 import app.igormatos.botaprarodar.BaseRobot
 import app.igormatos.botaprarodar.R
-import app.igormatos.botaprarodar.presentation.main.MainActivity
+import app.igormatos.botaprarodar.presentation.bikeForm.BikeFormActivity
+import app.igormatos.botaprarodar.presentation.welcome.WelcomeActivity
 import com.brunotmgomes.ui.extensions.REQUEST_PHOTO
+import org.hamcrest.Matcher
 import org.hamcrest.core.AllOf.allOf
 
 fun bicycle(executeFun: BicycleRobot.() -> Unit) = BicycleRobot().apply { executeFun() }
@@ -40,10 +45,6 @@ class BicycleRobot : BaseRobot() {
         fillFieldByHint("Número de ordem", content)
     }
 
-    fun fillBicyclePhoto() {
-        onView(withId(R.id.bikePhotoImageView)).perform()
-    }
-
     fun clickRegisterBicycle() {
         onView(withId(R.id.saveButton)).perform(click())
     }
@@ -58,19 +59,6 @@ class BicycleRobot : BaseRobot() {
 
     infix fun verify(executeFun: BicycleRobot.() -> Unit) {
         executeFun()
-    }
-
-    private fun createImageToActivityResult(): Instrumentation.ActivityResult? {
-        val bundle = Bundle()
-        bundle.putParcelable(
-            REQUEST_PHOTO.toString(), BitmapFactory.decodeResource(
-                IntentsTestRule(MainActivity::class.java).activity.resources,
-                R.drawable.ic_check_dialog
-            )
-        )
-        val resultData = Intent()
-        resultData.putExtras(bundle)
-        return Instrumentation.ActivityResult(Activity.RESULT_OK, resultData)
     }
 
 }
